Update CUDA Installer for 10.2
+ Add Automatic Plugin Key Generation for Yolo and TensorFlow plugins
+ Change gcc and g++ versions that are installed to version 8

+++ INSTALL/cuda.sh
@@ -3,42 +3,42 @@ echo "------------------------------------------"
 echo "-- Installing CUDA Toolkit and CUDA DNN --"
 echo "------------------------------------------"
 # Install CUDA Drivers and Toolkit
-echo "============="
-echo " Detecting Ubuntu Version"
-echo "============="
-getubuntuversion=$(lsb_release -r | awk '{print $2}' | cut -d . -f1)
-echo "============="
-echo " Ubuntu Version: $getubuntuversion"
-echo "============="
-if [ "$getubuntuversion" = "17" ] || [ "$getubuntuversion" > "17" ]; then
-    wget https://cdn.shinobi.video/installers/cuda-repo-ubuntu1710_9.2.148-1_amd64.deb -O cuda.deb
-    sudo apt-key adv --fetch-keys https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1710/x86_64/7fa2af80.pub
-    sudo dpkg -i cuda.deb
-fi
-if [ "$getubuntuversion" = "16" ]; then
-    wget https://cdn.shinobi.video/installers/cuda-repo-ubuntu1604_9.2.148-1_amd64.deb -O cuda.deb
-    sudo apt-key adv --fetch-keys https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1604/x86_64/7fa2af80.pub
-    sudo dpkg -i cuda.deb
-fi
-sudo apt-get update -y
-if [ "$getubuntuversion" = "17" ] || [ "$getubuntuversion" > "17" ]; then
+if [ -x "$(command -v apt)" ]; then
+    wget https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/cuda-ubuntu1804.pin
+    sudo mv cuda-ubuntu1804.pin /etc/apt/preferences.d/cuda-repository-pin-600
+    sudo apt-key adv --fetch-keys https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/7fa2af80.pub
+    sudo add-apt-repository "deb http://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/ /"
+    sudo apt-get update
+
+    sudo apt-get update -y
+
     sudo apt-get -o Dpkg::Options::="--force-overwrite" install cuda -y --no-install-recommends
     sudo apt-get -o Dpkg::Options::="--force-overwrite" install --fix-broken -y
+
+    # Install CUDA DNN
+    wget https://cdn.shinobi.video/installers/libcudnn7_7.6.5.32-1+cuda10.2_amd64.deb -O cuda-dnn.deb
+    sudo dpkg -i cuda-dnn.deb
+    wget https://cdn.shinobi.video/installers/libcudnn7-dev_7.6.5.32-1+cuda10.2_amd64.deb -O cuda-dnn-dev.deb
+    sudo dpkg -i cuda-dnn-dev.deb
+    echo "-- Cleaning Up --"
+    # Cleanup
+    sudo rm cuda-dnn.deb
+    sudo rm cuda-dnn-dev.deb
 fi
-if [ "$getubuntuversion" = "16" ]; then
-    sudo apt-get install libcuda1-384 -y --no-install-recommends
-    sudo apt-get install nvidia-cuda-toolkit -y
+if [ -x "$(command -v yum)" ]; then
+    sudo yum-config-manager --add-repo http://developer.download.nvidia.com/compute/cuda/repos/rhel7/x86_64/cuda-rhel7.repo
+    sudo yum clean all
+    sudo yum -y install nvidia-driver-latest-dkms cuda
+    sudo yum -y install cuda-drivers
+    wget https://cdn.shinobi.video/installers/libcudnn7-7.6.5.33-1.cuda10.2.x86_64.rpm -O cuda-dnn.rpm
+    sudo yum -y localinstall cuda-dnn.rpm
+    wget https://cdn.shinobi.video/installers/libcudnn7-devel-7.6.5.33-1.cuda10.2.x86_64.rpm -O cuda-dnn-dev.rpm
+    sudo yum -y localinstall cuda-dnn-dev.rpm
+    echo "-- Cleaning Up --"
+    sudo rm cuda-dnn.rpm
+    sudo rm cuda-dnn-dev.rpm
 fi
-# Install CUDA DNN
-wget https://cdn.shinobi.video/installers/libcudnn7_7.2.1.38-1+cuda9.2_amd64.deb -O cuda-dnn.deb
-sudo dpkg -i cuda-dnn.deb
-wget https://cdn.shinobi.video/installers/libcudnn7-dev_7.2.1.38-1+cuda9.2_amd64.deb -O cuda-dnn-dev.deb
-sudo dpkg -i cuda-dnn-dev.deb
-echo "-- Cleaning Up --"
-# Cleanup
-sudo rm cuda.deb
-sudo rm cuda-dnn.deb
-sudo rm cuda-dnn-dev.deb
+
 echo "------------------------------"
 echo "Reboot is required. Do it now?"
 echo "------------------------------"
